Make Hmac.prototype.update a clone of Hash.prototype.update.
Node's crypto.js just sets Hmac.prototype.update to be a reference to
Hash.prototype.update. Here this is changed to be a copy/paste job.
This is due to some strange behavior when running against nashorn where
Hmac.prototype.update mysteriously becomes a reference to
Hmac.prototype.update after the first iteration of a test. It's not
clear exactly why this is happening, and there is nothing in the code
which I can see that would cause this. Additionally, this behavior does
not occur with DynJS, so I suspect that there are issues with the
Nashorn linker. But again... not clear.